Rename CloudClient -> CDMClient
